+++ src/components/VideoPlayer/getVideoThumbnail.ts
@@ -1,10 +1,15 @@
 import { SizeValues } from "../../styles/utils/size";
 
-export const getVideoThumbnail = (
-  mainImage: string,
-  thumbTime: number | null = 20,
-  width: SizeValues = 400,
-  height: SizeValues = 200
-) => {
-  return `https://image.mux.com/${mainImage}/thumbnail.png?width=${width}&height=${height}&fit_mode=smartcrop&time=${thumbTime}`;
+type GetVideoThumbnailProps = {
+  video: { playbackId: string; thumbTime: number | null };
+  width?: SizeValues;
+  height?: SizeValues;
+};
+export const getVideoThumbnail = ({
+  video,
+  width = 400,
+  height = 200,
+}: GetVideoThumbnailProps) => {
+  const { playbackId, thumbTime = 200 } = video;
+  return `https://image.mux.com/${playbackId}/thumbnail.png?width=${width}&height=${height}&fit_mode=smartcrop&time=${thumbTime}`;
 };