+++ src/components/pages/WebinarsIndex.page.tsx
@@ -8,6 +8,7 @@ import {
 } from "../../common-lib/cms-types";
 import { WebinarsListingPage } from "../../common-lib/cms-types/webinarsListingPage";
 import PostListing from "../Posts/PostListing";
+import { getVideoThumbnail } from "../VideoPlayer/getVideoThumbnail";
 
 export type SerializedWebinarPreview = Omit<WebinarPreview, "date"> & {
   date: string;
@@ -56,8 +57,7 @@ export const webinarToBlogListItem = (
   titleTag: "h3",
   category: webinar.category,
   date: webinar.date,
-  mainImage: webinar.video.video.asset.playbackId,
-  thumbTime: webinar.video.video.asset.thumbTime,
+  thumbnailUrl: getVideoThumbnail({ video: webinar.video.video.asset }),
 });
 
 export default WebinarListingPage;